Make conversation in different tab
