+++ src/App.js
@@ -8,6 +8,7 @@ import './sass/app.css';
 import { getMessages, setMessages } from './store/chat/chat';
 import LogIn from './components/logIn';
 import SignUp from './components/signUp';
+import Conversation from './components/chats/chatConversation';
 
 
 function App() {
@@ -25,7 +26,8 @@ function App() {
         <Routes>
           <Route path="/" element={<Navbar/>}/>
           <Route path="/login" element={<LogIn/>}/>
-          <Route path="/Home" element={<Home/>}/>
+          <Route path="/Home" exact element={<Home/>}/>
+          <Route path="/conversation/:id" element={<Conversation/>}/>
           <Route path="/signup" element={<SignUp/>}/>
         </Routes>
